+++ src/fights/competitions.clj
@@ -11,10 +11,12 @@
         :body)))
 
 (defn insert-competitions
+  "cometitions テーブルを初期化し、
+   ijf の competions リストから
+   id_competion, comp_year, name, has_results を読み込む。"
   []
   (db/create-competitions)
   (doseq [c (get-competitions)]
-    (println c)
+    ;;(println c)
     (db/insert-competition c)))
 
-(insert-competitions)