+++ test/fights/db_test.clj
@@ -8,5 +8,5 @@
     (let [ret (insert-competition {:name "name"
                                    :has_results 10
                                    :comp_year 2022
-                                   :id_competition -1})]
-      (println "test ret=" ret))))
+                                   :id_competition -4})]
+      (println "ret" ret))))